Version Page and Page#content with reverting
PT-AT was generating even more queries and wasn't terribly helpeful in
presenting a way to get to the object_changes. If I tried to reduce the
`touch`-ing queries, it would break PT-AT anyways so I decided to
simplify my dependencies and treat reverting the page's attributes vs
the page's content as separate.