Skip auth check on /rails/* routes
File uploads for Trix/ActionText would fail due to AUTH_TOKEN cookie not
being sent with the uploaded document. Due to the CSRF token that is
being sent, and to get a CSRF token, you need to first load the /new
page, I believe this is secure in the interim until I can figure out how
to inject the cookie into the Trix.js file uploads.